requirements.txt
fastapi
uvicorn[standard]
line-bot-sdk
google.generativeai
grpcio
pillow
requests
git+https://github.com/ozgur/python-firebase
